requirements.txt
fastapi
uvicorn
qdrant-client
openai
langchain
cassandra-driver
datasets
